Data & Automation Agents/API response interpreter agent/requirements.txt
openai
litellm
streamlit
python-dotenv
pandas
google-generativeai
anthropic
